change JulianDay type from int to float; add more date specification properties

+++ src/wisski_pathbuilder/Pathbuilder.py
@@ -150,7 +150,7 @@ class WisskiPath:
         elif valuetype == 'xsd:coordinates':
             fieldargs = ('geofield', 'geofield_latlon', 'geofield_latlon')
         elif valuetype == 'spec:JulianDay':
-            fieldargs = ('integer', 'number', 'number_unformatted')
+            fieldargs = ('float', 'number', 'number_unformatted')
         elif valuetype != 'xsd:string':
             warn(f'Unrecognised datatype {valuetype}; setting as string')
         self.fieldtype = fieldargs[0]
